unify repeating code
