+++ src/libraries/Native/Unix/System.Native/pal_errno.c
@@ -6,205 +6,13 @@
 #include "pal_errno.h"
 #include "pal_utilities.h"
 
-#include <errno.h>
 #include <netdb.h>
-
-// ENODATA is not defined in FreeBSD 10.3 but is defined in 11.0
-#if defined(__FreeBSD__) & !defined(ENODATA)
-#define ENODATA ENOATTR
-#endif
-
 #include <string.h>
 #include <assert.h>
 
 int32_t SystemNative_ConvertErrorPlatformToPal(int32_t platformErrno)
 {
-    switch (platformErrno)
-    {
-        case 0:
-            return Error_SUCCESS;
-        case E2BIG:
-            return Error_E2BIG;
-        case EACCES:
-            return Error_EACCES;
-        case EADDRINUSE:
-            return Error_EADDRINUSE;
-        case EADDRNOTAVAIL:
-            return Error_EADDRNOTAVAIL;
-        case EAFNOSUPPORT:
-            return Error_EAFNOSUPPORT;
-        case EAGAIN:
-            return Error_EAGAIN;
-        case EALREADY:
-            return Error_EALREADY;
-        case EBADF:
-            return Error_EBADF;
-        case EBADMSG:
-            return Error_EBADMSG;
-        case EBUSY:
-            return Error_EBUSY;
-        case ECANCELED:
-            return Error_ECANCELED;
-        case ECHILD:
-            return Error_ECHILD;
-        case ECONNABORTED:
-            return Error_ECONNABORTED;
-        case ECONNREFUSED:
-            return Error_ECONNREFUSED;
-        case ECONNRESET:
-            return Error_ECONNRESET;
-        case EDEADLK:
-            return Error_EDEADLK;
-        case EDESTADDRREQ:
-            return Error_EDESTADDRREQ;
-        case EDOM:
-            return Error_EDOM;
-        case EDQUOT:
-            return Error_EDQUOT;
-        case EEXIST:
-            return Error_EEXIST;
-        case EFAULT:
-            return Error_EFAULT;
-        case EFBIG:
-            return Error_EFBIG;
-        case EHOSTUNREACH:
-            return Error_EHOSTUNREACH;
-        case EIDRM:
-            return Error_EIDRM;
-        case EILSEQ:
-            return Error_EILSEQ;
-        case EINPROGRESS:
-            return Error_EINPROGRESS;
-        case EINTR:
-            return Error_EINTR;
-        case EINVAL:
-            return Error_EINVAL;
-        case EIO:
-            return Error_EIO;
-        case EISCONN:
-            return Error_EISCONN;
-        case EISDIR:
-            return Error_EISDIR;
-        case ELOOP:
-            return Error_ELOOP;
-        case EMFILE:
-            return Error_EMFILE;
-        case EMLINK:
-            return Error_EMLINK;
-        case EMSGSIZE:
-            return Error_EMSGSIZE;
-        case EMULTIHOP:
-            return Error_EMULTIHOP;
-        case ENAMETOOLONG:
-            return Error_ENAMETOOLONG;
-        case ENETDOWN:
-            return Error_ENETDOWN;
-        case ENETRESET:
-            return Error_ENETRESET;
-        case ENETUNREACH:
-            return Error_ENETUNREACH;
-        case ENFILE:
-            return Error_ENFILE;
-        case ENOBUFS:
-            return Error_ENOBUFS;
-        case ENODEV:
-            return Error_ENODEV;
-        case ENOENT:
-            return Error_ENOENT;
-        case ENOEXEC:
-            return Error_ENOEXEC;
-        case ENOLCK:
-            return Error_ENOLCK;
-        case ENOLINK:
-            return Error_ENOLINK;
-        case ENOMEM:
-            return Error_ENOMEM;
-        case ENOMSG:
-            return Error_ENOMSG;
-        case ENOPROTOOPT:
-            return Error_ENOPROTOOPT;
-        case ENOSPC:
-            return Error_ENOSPC;
-        case ENOSYS:
-            return Error_ENOSYS;
-        case ENOTCONN:
-            return Error_ENOTCONN;
-        case ENOTDIR:
-            return Error_ENOTDIR;
-#if ENOTEMPTY != EEXIST // AIX defines this
-        case ENOTEMPTY:
-            return Error_ENOTEMPTY;
-#endif
-#ifdef ENOTRECOVERABLE // not available in NetBSD
-        case ENOTRECOVERABLE:
-            return Error_ENOTRECOVERABLE;
-#endif
-        case ENOTSOCK:
-            return Error_ENOTSOCK;
-        case ENOTSUP:
-            return Error_ENOTSUP;
-        case ENOTTY:
-            return Error_ENOTTY;
-        case ENXIO:
-            return Error_ENXIO;
-        case EOVERFLOW:
-            return Error_EOVERFLOW;
-#ifdef EOWNERDEAD // not available in NetBSD
-        case EOWNERDEAD:
-            return Error_EOWNERDEAD;
-#endif
-        case EPERM:
-            return Error_EPERM;
-        case EPIPE:
-            return Error_EPIPE;
-        case EPROTO:
-            return Error_EPROTO;
-        case EPROTONOSUPPORT:
-            return Error_EPROTONOSUPPORT;
-        case EPROTOTYPE:
-            return Error_EPROTOTYPE;
-        case ERANGE:
-            return Error_ERANGE;
-        case EROFS:
-            return Error_EROFS;
-        case ESPIPE:
-            return Error_ESPIPE;
-        case ESRCH:
-            return Error_ESRCH;
-        case ESTALE:
-            return Error_ESTALE;
-        case ETIMEDOUT:
-            return Error_ETIMEDOUT;
-        case ETXTBSY:
-            return Error_ETXTBSY;
-        case EXDEV:
-            return Error_EXDEV;
-#ifdef ESOCKTNOSUPPORT
-        case ESOCKTNOSUPPORT:
-            return Error_ESOCKTNOSUPPORT;
-#endif
-        case EPFNOSUPPORT:
-            return Error_EPFNOSUPPORT;
-        case ESHUTDOWN:
-            return Error_ESHUTDOWN;
-        case EHOSTDOWN:
-            return Error_EHOSTDOWN;
-        case ENODATA:
-            return Error_ENODATA;
-
-// #if because these will trigger duplicate case label warnings when
-// they have the same value, which is permitted by POSIX and common.
-#if EOPNOTSUPP != ENOTSUP
-        case EOPNOTSUPP:
-            return Error_EOPNOTSUPP;
-#endif
-#if EWOULDBLOCK != EAGAIN
-        case EWOULDBLOCK:
-            return Error_EWOULDBLOCK;
-#endif
-    }
-
-    return Error_ENONSTANDARD;
+    return ConvertErrorPlatformToPal(platformErrno);
 }
 
 int32_t SystemNative_ConvertErrorPalToPlatform(int32_t error)
